src/style.css: remove this
As far as I can tell this doesn't have any CSS settings that are specific to
the UAT chooser. On my browser, it does make the hyperlinks blue -- if I get
rid of this file the hyperlinks are black. That's a bit odd to me since I
thought everyone's default was blue hyperlinks, but, well, whatever.