src/ejp_uat.js: add href="javascript:void(0);" to generate <a> tags
Without an `href` attribute, browsers won't style them as active links. This
can be worked around in CSS, but it means that users of the widget are going
to need to implement some custom CSS, which we'd like to avoid. Based on the
relevant StackOverflow discussion
(https://stackoverflow.com/questions/4855168/what-is-href-and-why-is-it-used),
it looks like the best solution is to use this sort of `javascript:` href. If
the widget user wants to customize the styling of these links with CSS, they
still can, of course, but this should yield nicer behavior by default.